Resizing subcomponents from context menu
resizing through the context menu in better shape. how does it work? for buttons and other juce componets, the buttonClicked handles this. For the rest, aka the panels, this is handled in panelbase::mouseDown. For resizing components, the owner will resize the component and the component will resize the subcomponents.